Add files via upload
The Python Stone-Paper-Scissors game is a classic hand game implemented in Python. It allows two players to compete against each other, with each player choosing one of the three options: stone, paper, or scissors. The game determines the winner based on the choices made by the players.  But you must remember the second player is none other than the system itself.